#394's branch gained six commits (0cf4fc0..5fd5994: signing oracles,
dead-stack residue, tightened tools/ gates, docs) after #403 was cut from
dcc8ab9. #403 then conflicted with its base. Resolved:

- static-analysis.yml: both sides fixed the Valgrind "7/6" count. Kept the
  base's form (`set -- $tests; expected=$#`) and its comment, plus #403's
  codeql-action v4.38.1 pins and the build/setuptools/wheel/cmake/Cython
  floors. The base's -Wsign-conversion lanes and the digest-pinned
  manylinux image come through unchanged.
- CHANGELOG.md: both sets of [Unreleased] entries kept.
- docs/METRICS_REPORT.md: regenerated with tools/update_docs.py --loc
  after staging, not hand-merged. (Measured mid-merge it counted the
  conflicted index stages twice.)

Validated on the merged tree:
- cmake Release build with no warnings; ctest 146/146 (ISA-gated
  kernels skipped on this CPU).
- check_documented_counts OK (67 across 14 families); check_action_pins
  --strict 142/142; check_workflow_commands, check_headers,
  check_docker_pins, check_version_consistency and check_documented_extras
  pass. ruff check and black --check are clean.
- 3,340 gate/workflow/docs/key-format tests pass. The 3 that fail
  (binding-extension artefact, two trust-anchor attestation tests) fail
  identically on the unmodified base: this machine builds without the
  Cython extensions that the signed artefact binds.

+++ src/c/ama_aes_gcm.c
@@ -186,12 +186,12 @@ static void aes256_encrypt_block(const uint8_t round_keys[240],
     int round;
 
     /* AddRoundKey (round 0) */
-    for (int j = 0; j < 16; j++)
+    for (size_t j = 0; j < 16; j++)
         state[j] = in[j] ^ round_keys[j];
 
     for (round = 1; round <= 14; round++) {
         /* SubBytes */
-        for (int j = 0; j < 16; j++)
+        for (size_t j = 0; j < 16; j++)
             t[j] = aes_sbox[state[j]];
 
         /* ShiftRows */
@@ -216,7 +216,7 @@ static void aes256_encrypt_block(const uint8_t round_keys[240],
 
         /* AddRoundKey */
         const uint8_t *rk = round_keys + round * 16;
-        for (int j = 0; j < 16; j++)
+        for (size_t j = 0; j < 16; j++)
             state[j] ^= rk[j];
     }
 
@@ -408,7 +408,7 @@ static void ghash(const uint8_t H[16],
     /* Process AAD */
     full_blocks = aad_len / 16;
     for (i = 0; i < full_blocks; i++) {
-        for (int j = 0; j < 16; j++)
+        for (size_t j = 0; j < 16; j++)
             S[j] ^= aad[i * 16 + j];
         ghash_mul(S, H);
     }
@@ -419,7 +419,7 @@ static void ghash(const uint8_t H[16],
          * writes on the secure path. */
         ama_secure_memzero(block, 16);
         memcpy(block, aad + full_blocks * 16, remaining);
-        for (int j = 0; j < 16; j++)
+        for (size_t j = 0; j < 16; j++)
             S[j] ^= block[j];
         ghash_mul(S, H);
     }
@@ -427,7 +427,7 @@ static void ghash(const uint8_t H[16],
     /* Process ciphertext */
     full_blocks = ct_len / 16;
     for (i = 0; i < full_blocks; i++) {
-        for (int j = 0; j < 16; j++)
+        for (size_t j = 0; j < 16; j++)
             S[j] ^= ciphertext[i * 16 + j];
         ghash_mul(S, H);
     }
@@ -435,7 +435,7 @@ static void ghash(const uint8_t H[16],
     if (remaining > 0) {
         ama_secure_memzero(block, 16);
         memcpy(block, ciphertext + full_blocks * 16, remaining);
-        for (int j = 0; j < 16; j++)
+        for (size_t j = 0; j < 16; j++)
             S[j] ^= block[j];
         ghash_mul(S, H);
     }
@@ -464,7 +464,7 @@ static void ghash(const uint8_t H[16],
         block[14] = (uint8_t)(ct_bits >> 8);
         block[15] = (uint8_t)(ct_bits);
     }
-    for (int j = 0; j < 16; j++)
+    for (size_t j = 0; j < 16; j++)
         S[j] ^= block[j];
     ghash_mul(S, H);
 
@@ -601,7 +601,7 @@ ama_error_t ama_aes256_gcm_encrypt(
     for (i = 0; i < full_blocks; i++) {
         gcm_inc32(counter);
         aes256_encrypt_block(round_keys, counter, keystream);
-        for (int j = 0; j < 16; j++)
+        for (size_t j = 0; j < 16; j++)
             ciphertext[i * 16 + j] = plaintext[i * 16 + j] ^ keystream[j];
     }
     remaining = pt_len % 16;
@@ -616,7 +616,7 @@ ama_error_t ama_aes256_gcm_encrypt(
     ghash(H, aad, aad_len, ciphertext, pt_len, tag);
 
     /* Final tag = GHASH XOR E_K(J0) */
-    for (int j = 0; j < 16; j++)
+    for (size_t j = 0; j < 16; j++)
         tag[j] ^= tag_mask[j];
 
     /* Scrub sensitive material */
@@ -718,7 +718,7 @@ ama_error_t ama_aes256_gcm_decrypt(
 
     /* Compute GHASH over AAD and ciphertext BEFORE decrypting */
     ghash(H, aad, aad_len, ciphertext, ct_len, computed_tag);
-    for (int j = 0; j < 16; j++)
+    for (size_t j = 0; j < 16; j++)
         computed_tag[j] ^= tag_mask[j];
 
     /* Constant-time tag comparison + unified post-verify control flow.
@@ -746,7 +746,7 @@ ama_error_t ama_aes256_gcm_decrypt(
     for (i = 0; i < bounded_full; i++) {
         gcm_inc32(counter);
         aes256_encrypt_block(round_keys, counter, keystream);
-        for (int j = 0; j < 16; j++)
+        for (size_t j = 0; j < 16; j++)
             plaintext[i * 16 + j] = ciphertext[i * 16 + j] ^ keystream[j];
     }
     if (bounded_remaining > 0) {
